Can view Members miles and delete members. Can search, update and delete accountTypes. tried to update miles but encounter errors

+++ AccountSystem/AccountSystemLogic/src/main/java/za/ac/nwu/ac/logic/flow/FetchMemberFlow.java
@@ -8,5 +8,5 @@ public interface FetchMemberFlow {
 
     List<MemberDto> getAllMembers();
 
-    MemberDto getMemberByContactNumber(int contactNumber);
+    MemberDto getMemberByContactNumber(long contactNumber);
 }
