+++ AccountSystem/AccountSystemLogic/src/main/java/za/ac/nwu/ac/logic/flow/impl/FetchMemberFlowImpl.java
@@ -28,7 +28,7 @@ public class FetchMemberFlowImpl implements FetchMemberFlow {
     }
 
     @Override
-    public MemberDto getMemberByContactNumber(int contactNumber){
+    public MemberDto getMemberByContactNumber(long contactNumber){
         return memberTranslator.getMemberByContactNumber(contactNumber);
     }
 }
